+++ src/test/kotlin/com/cmgapps/license/reporter/JsonReportShould.kt
@@ -59,8 +59,8 @@ class JsonReportShould {
                         ]
                     }
                 ]
-                """.trimIndent()
-            )
+                """.trimIndent(),
+            ),
         )
     }
 }