+++ src/test/kotlin/com/cmgapps/license/util/LibrariesHelper.kt
@@ -42,5 +42,5 @@ internal val testLibraries = listOf(
                 url = "https://www.apache.org/licenses/LICENSE-2.0.txt",
             ),
         ),
-    )
+    ),
 )